put the date of the next ceremony in the right field and I'm pretty sure we can now merge it.
