+++ src/map/MapCeremonyPhases.jsx
@@ -1,6 +1,9 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { Step, Label } from 'semantic-ui-react';
 import { CeremonyPhaseTimer } from './CeremonyPhaseTimer';
+import { getNextMeetupTime } from '@encointer/node-api';
+import { stringToDegree } from '@encointer/types';
+import { bnToU8a } from '@polkadot/util';
 
 const ceremonyPhases = [
   'REGISTERING',
@@ -12,12 +15,36 @@ const formatDate = (timestamp) => (new Date(timestamp)).toLocaleString();
 
 const formatStartingAt = (timestamp) => (<div><div>starting at:</div> {formatDate(timestamp)}</div>);
 
+/**
+ * Parses a location json with fields as number strings to a `Location` object.
+ *
+ * There is a rust vs. JS endian issue with numbers: https://github.com/polkadot-js/api/issues/4313.
+ *
+ * tl;dr: If the returned location is processed:
+ *  * by a node (rust), use isLe = false.
+ *  * by JS, e.g. `parseDegree`, use isLe = true.
+ *
+ *
+ * @param api
+ * @param location fields as strings, e.g. '35.2313515312'
+ * @param isLe
+ * @returns {Location} Location with fields as fixed-point numbers
+ */
+export function locationFromJson (api, location, isLe = true) {
+  return api.createType('Location', {
+    lat: bnToU8a(stringToDegree(location.lat), 128, isLe),
+    lon: bnToU8a(stringToDegree(location.lon), 128, isLe)
+  });
+}
+
 export default React.memo(function MapCeremonyPhases (props) {
   const {
     small,
     participantCount,
     meetupCount,
     attestationCount,
+    api,
+    cids,
     currentPhase: {
       phase: currentPhase,
       timestamp
@@ -38,6 +65,21 @@ export default React.memo(function MapCeremonyPhases (props) {
       );
     return (small && currentPhase === 2) ? phasesProps.reverse() : phasesProps;
   };
+
+  const [nextMeetupTime, setNextMeetupTime] = useState([]);
+  // gets the date of the next Meetup
+  useEffect(() => {
+    async function getNextMeetupDate () {
+      const meetupLocations = await api.rpc.encointer.getLocations(cids[0]);
+      const tempLocation = locationFromJson(api, meetupLocations[0]);
+      const tempTime = await getNextMeetupTime(api, tempLocation);
+      const tempdate = new Date(tempTime.toNumber());
+      const temparray = (tempdate.toString()).split('G');
+      const tempresultarray = temparray[0].split(' ');
+      setNextMeetupTime(' ' + tempresultarray[0] + ' ' + tempresultarray[2] + ' ' + tempresultarray[1] + ' ' + tempresultarray[3]);
+    }
+    getNextMeetupDate();
+  }, [api, nextMeetupTime, setNextMeetupTime, cids]);
 
   return (<div className='encointer-map-ceremony-phase'>
     <Step.Group
@@ -66,6 +108,13 @@ export default React.memo(function MapCeremonyPhases (props) {
                             ? formatStartingAt(timestamp)
                             : null))
                 }</Step.Description>
+                <Step.Description>
+                  {
+                    (props.key === 'ATTESTING')
+                      ? <div>date of next ceremony: {nextMeetupTime}</div>
+                      : null
+                  }
+                </Step.Description>
               </Step.Content>
             </Step>
           ))
